removed main activity and login button done
