+++ app/src/main/res/layout/activity_login.xml
@@ -132,7 +132,7 @@
                 android:layout_marginLeft="24dp"
                 android:layout_marginTop="8dp"
                 android:layout_marginRight="24dp"
-                android:layout_marginBottom="16dp"
+                android:layout_marginBottom="50dp"
                 android:paddingBottom="5dp"
                 android:text="@string/login" />
